+++ corrupted_clock/src/table_drawing.rs
@@ -1,5 +1,9 @@
-use corrupted_clock_util::timing::{CountDown, Stopwatch, TimeImpl, Timer};
+use corrupted_clock_util::timing::{
+    ClockDuration, CountDown, Stopwatch, TimeImpl, Timer, UtcDateTime,
+};
 use prettytable::{Cell, Row, Table};
+
+use crate::constants::NOT_AVIABLE_TXT;
 
 pub fn count_down_rows<'a, T>(
     count_downs: impl IntoIterator<Item = (&'a str, &'a CountDown<T>)>,
@@ -25,13 +29,20 @@ pub fn stop_watch_rows<'a, T>(
 where
     T: Default + TimeImpl + 'a,
 {
+    type CellVec = Vec<Cell>;
     let stop_watches = count_downs.into_iter();
 
     let mut table = Table::new();
-    table.add_row(Row::new(stopwatch_header().collect()));
+    let headers: CellVec = stopwatch_header().collect();
+    let elements_num = headers.len();
+    table.add_row(Row::new(headers));
 
     for name_stop_watch in stop_watches {
-        table.add_row(stopwatch_fields(name_stop_watch).collect());
+        let fields: CellVec = stopwatch_fields(name_stop_watch)
+            .map(|e| Cell::new(&e))
+            .collect();
+        debug_assert_eq!(elements_num, fields.len());
+        table.add_row(fields.into());
     }
     table
 }
@@ -40,31 +51,53 @@ fn stopwatch_header() -> impl Iterator<Item = Cell> {
     [
         "Name",
         "Created at",
+        "Started at",
         "Is paused",
         "Passed Time",
         "Paused Time",
+        "Last resumed at",
+        "Last paused at",
     ]
     .map(Cell::new)
     .into_iter()
-}
-
-fn count_down_headers() -> impl Iterator<Item = Cell> {
-    ["Count down", "Left Time"].map(Cell::new).into_iter()
 }
 
 fn stopwatch_fields<'a, T>((name, stop_watch): (&'a str, &'a T)) -> impl Iterator<Item = String>
 where
     T: Timer + 'a,
 {
-    let local_create_at = corrupted_clock_util::convert_utc_to_local(stop_watch.created_at());
+    fn to_local_short_table_field(date: UtcDateTime) -> String {
+        corrupted_clock_util::chrono_time_to_str(corrupted_clock_util::convert_utc_to_local(date))
+    }
+
+    fn convert_to_opt_table_field(opt: Option<UtcDateTime>) -> String {
+        if let Some(date) = opt {
+            to_local_short_table_field(date)
+        } else {
+            NOT_AVIABLE_TXT.to_string()
+        }
+    }
+
+    let local_create_at = to_local_short_table_field(stop_watch.created_at());
+    let local_started_at = to_local_short_table_field(stop_watch.start_moment());
+    let last_resumed = convert_to_opt_table_field(stop_watch.last_resumed_at());
+    let last_paused = convert_to_opt_table_field(stop_watch.last_paused_at());
+
     [
         name.to_string(),
-        local_create_at.to_string(),
+        local_create_at,
+        local_started_at,
         stop_watch.is_paused().to_string(),
         stop_watch.passed().to_string(),
         stop_watch.paused_time().to_string(),
+        last_resumed,
+        last_paused,
     ]
     .into_iter()
+}
+
+fn count_down_headers() -> impl Iterator<Item = Cell> {
+    ["Count down", "Left Time"].map(Cell::new).into_iter()
 }
 
 fn count_down_fields<'a, T>(cd: &CountDown<T>) -> impl Iterator<Item = String>